+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1507,7 +1507,7 @@ class Evaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-29T16:31:49.666Z")
+    ] = parse_datetime("2026-06-30T03:15:25.282Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1652,7 +1652,7 @@ class Evaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-29T16:31:49.665Z")
+    ] = parse_datetime("2026-06-30T03:15:25.281Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1797,7 +1797,7 @@ class Evaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-29T16:31:49.664Z")
+    ] = parse_datetime("2026-06-30T03:15:25.280Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1902,7 +1902,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-29T16:31:22.968Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-30T03:15:02.428Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")