+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWA3J1T3Q7YD27Y7MDBX1FHC"
+        "01KWB8CGC9WCBQ9YEPZ4CAQX83"
     )
     r"""The unique identifier of the data source"""