+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWA3J1T8CC1N0GPDVVFGQCJF"
+        "01KWB8CGCD8RCHKTHM4E3ZJBYA"
     )
     r"""The unique identifier of the data source"""