+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KWA3J1P3TGWJ3MDZP6C83E0V"
+    id: Optional[str] = "01KWB8CG8EGFM7P4QKH25V5CBM"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWA3J1NVCRTFSRF1B9S22B90"
+        "tool_01KWB8CG86EC6TP29XSAXRPB67"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KWA3J1NTAEJZFG3VQCDNFXDP"
+    id: Optional[str] = "01KWB8CG85RSMPM0PZXC7ZGTG2"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWA3J1NSE16DPS5AD93M6PQ7"
+        "tool_01KWB8CG840DTMHJX87YXFP94M"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWA3J1NPG5HKHCS3MTJ37BZ3"
+        "tool_01KWB8CG82597Y32SS2W98MNWE"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWA3J1NMAJ5J36H1E5C7ZV3K"
+        "tool_01KWB8CG80SC95K1D2AB978F4G"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWA3J1NKQWANYQD755WAWGGV"
+        "tool_01KWB8CG7YCBAK5D9M6ZDWAGG5"
     )
 
     display_name: Optional[str] = None